Metadata fetch using Oauth token UI related change
