+++ src/main/webapp/css/homepage/hometheme.css
@@ -1,0 +1,22 @@
+.w3-theme-l5 {color:#000 !important; background-color:#f0f0f0 !important}
+.w3-theme-l4 {color:#000 !important; background-color:#cccccc !important}
+.w3-theme-l3 {color:#fff !important; background-color:#999999 !important}
+.w3-theme-l2 {color:#fff !important; background-color:#666666 !important}
+.w3-theme-l1 {color:#fff !important; background-color:#333333 !important}
+.w3-theme-d1 {color:#fff !important; background-color:#000000 !important}
+.w3-theme-d2 {color:#fff !important; background-color:#000000 !important}
+.w3-theme-d3 {color:#fff !important; background-color:#000000 !important}
+.w3-theme-d4 {color:#fff !important; background-color:#000000 !important}
+.w3-theme-d5 {color:#fff !important; background-color:#000000 !important}
+
+.w3-theme-light {color:#000 !important; background-color:#f0f0f0 !important}
+.w3-theme-dark {color:#fff !important; background-color:#000000 !important}
+.w3-theme-action {color:#fff !important; background-color:#000000 !important}
+
+.w3-theme {color:#fff !important; background-color:#000000 !important}
+.w3-text-theme {color:#000000 !important}
+.w3-border-theme {border-color:#000000 !important}
+
+.w3-hover-theme:hover {color:#000 !important; background-color:#ffc107 !important}
+.w3-hover-text-theme:hover {color:#000000 !important}
+.w3-hover-border-theme:hover {border-color:#000000 !important}
